+++ R/leaflet.extras-package.r
@@ -1,0 +1,6 @@
+#' leaflet.extras.
+#' @importFrom magrittr %>%
+#'
+#' @name leaflet.extras
+#' @docType package
+NULL