+++ NEWS.md
@@ -1,5 +1,7 @@
 # leaflet.extras 0.1.0
 
+* Added pulse-icon plugin
+
 * Added a `NEWS.md` file to track changes to the package.